requirements.txt
flask==3.0.0
flask-cors==4.0.0
python-dotenv==1.0.0
openai==1.54.3
requests==2.31.0
beautifulsoup4==4.12.0

# Google Drive Integration
google-api-python-client==2.108.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.2.0

# LangChain for Universal Document Loading
langchain==0.1.0
langchain-community==0.0.10
unstructured==0.11.6
unstructured[all-docs]==0.11.6

# Enhanced Document Processing
PyMuPDF==1.23.8
python-docx==1.1.0
Pillow==10.1.0
pytesseract==0.3.10
pypdf==3.17.4
pdfminer.six==20221105

# Excel & CSV
openpyxl==3.1.2
pandas==2.1.4
xlrd==2.0.1

# PowerPoint
python-pptx==0.6.23

# Markdown & HTML
markdown==3.5.1
html2text==2020.1.16

# Search & Indexing
whoosh==2.7.4

# Additional utilities
chardet==5.2.0
python-magic==0.4.27